move function to projects.go and add comments
