+++ projects.go
@@ -207,3 +207,38 @@ func (g *Gitlab) ProjectMergeRequests(id string, page int, per_page int, state s
 
 	return mr, err
 }
+
+/*
+Get single project id.
+
+    GET /projects/search/:query
+
+Parameters:
+
+    namespace The namespace of a project
+    name      The id of a project
+
+*/
+func (g *Gitlab) SearchProjectId(namespace string, name string) (id int, err error) {
+
+	url, opaque := g.ResourceUrlRaw(project_url_search_ids, map[string]string{
+		":query": name,
+	})
+
+	var projects []*Project
+
+	contents, err := g.buildAndExecRequestRaw("GET", url, opaque, nil)
+	if err == nil {
+		err = json.Unmarshal(contents, &projects)
+	} else {
+		return id, err
+	}
+
+	for _, project := range projects {
+		if project.Namespace.Name == namespace {
+			id = project.Id
+		}
+	}
+
+	return id, err
+}
